Struct mailstrom::internal_status::InternalStatus

pub struct InternalStatus {
    pub message_id: String,
    pub recipients: Vec<Recipient>,
    pub delivered_to_mx: Vec<SocketAddr>,
}

An email to be sent (internal format). This is exposed publicly for implementers of MailstromStorage

Fields
message_id: String

The parsed-out (or generated) message ID

recipients: Vec<Recipient>

The parsed-out list of recipients, and the state each is in. If this is None, then the recipient information has not been determined yet (MX record lookups take some time).

delivered_to_mx: Vec<SocketAddr>

The MX servers which we have successfully delivered to. We keep this as a separate list so we can more easily detect the situation where multiple recipients deliveries go to the same MX

Methods

impl InternalStatus

fn create(email: Email,
helo_name: &str)
-> Result<(InternalStatus, Email), Error>

fn as_status(&self) -> Status