Struct mailstrom::Mailstrom

pub struct Mailstrom<S: MailstromStorage + 'static> { /* fields omitted */ }

Methods

impl<S: MailstromStorage + 'static> Mailstrom<S>

fn new(config: Config, storage: S) -> Mailstrom<S>

Create a new Mailstrom instance for sending emails.

fn die(&mut self) -> Result<(), Error>

Ask Mailstrom to die. This is not required, you can simply let it fall out of scope and it will clean itself up.

fn worker_status(&self) -> WorkerStatus

Determine the status of the worker

fn send_email(&mut self, email: Email) -> Result<String, Error>

Send an email, getting back it's message-id

fn query_status(&mut self, message_id: &str) -> Result<Status, Error>